Get player counting stats from hockey-reference.com

Get player counting stats from hockey-reference.com

get_player_stats_hr(player_name, season = "career", league = "NHL")

Arguments

player_name
A player name or vector of player names

season
An integer value denoting the end year of the season(s) to scrape

league
The league stats to scrape, either 'NHL' or 'WHA'

Value

A tibble containing goals, assists, and various other stats for the specified player(s) from hockey-reference.com

Examples

if (FALSE) {
#' get_player_stats("Wayne Gretzky")
get_player_stats_hr(c("Wayne Gretzky","Mario Lemieux"))
}